requirements.txt
requests
beautifulsoup4
Pillow
ddddocr
loguru